Update: Event-Driven UI Refresh
UI에서 로직을 구현할 때, 상속 관계가 아닌 것(느슨한 결합)의 상태를 변경하는 로직을 Scriptableobject-Architecture의 GameEvent 사용하기

- 실제로 적용한 부분
상점이나 소환에서 유닛 혹은 아이템을 뽑으면 이를 유닛 정보나 아티팩트, 고서 정보에서 최신화하기

+++ Assets/EvolveThisMatch/Script/Lobby/AgentList/AgentInfo/UIAgentListItem_AgentInfo.cs
@@ -85,7 +85,7 @@ namespace EvolveThisMatch.Lobby
             else
             {
                 _counterText.text = $"{unitCount}/{maxUnitCount}";
-                _counterImage.fillAmount = unitCount / maxUnitCount;
+                _counterImage.fillAmount = unitCount / (float)maxUnitCount;
             }
 
             _dim.enabled = false;
